Codechange: Use ProviderManager to define screenshot generators.
