+++ src/CMakeLists.txt
@@ -40,6 +40,11 @@ add_files(
 )
 
 add_files(
+    screenshot_png.cpp
+    CONDITION PNG_FOUND
+)
+
+add_files(
     soundloader_opus.cpp
     CONDITION OpusFile_FOUND
 )
@@ -381,10 +386,13 @@ add_files(
     roadveh_cmd.h
     roadveh_gui.cpp
     safeguards.h
+    screenshot_bmp.cpp
     screenshot_gui.cpp
     screenshot_gui.h
     screenshot.cpp
     screenshot.h
+    screenshot_type.h
+    screenshot_pcx.cpp
     settings.cpp
     settings_cmd.h
     settings_func.h
